# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: BlindWay REST API
  version: 1.0.0
  description: 手机端、管理端与EMQX使用的BlindWay后端契约。所有手机坐标均为WGS84，经度在前、纬度在后。
  license:
    name: Apache License 2.0
    identifier: Apache-2.0
servers:
  - url: /api/v1
    description: 与当前文档站点同源的API
...
paths:
...
  /trips:
    post:
      tags: [Trips]
      operationId: startTrip
      summary: 为已绑定设备开始行程
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StartTripRequest' }
      responses:
        '201':
          description: 行程已开始
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Trip' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /trips/{tripId}/track-points:
    post:
      tags: [Trips]
      operationId: appendTrackPoints
      summary: 批量追加手机WGS84轨迹点
      parameters:
        - $ref: '#/components/parameters/TripId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TrackPointBatchRequest' }
      responses:
        '202': { description: 轨迹点已接受 }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /trips/{tripId}/completion:
    post:
      tags: [Trips]
      operationId: completeTrip
      summary: 完成进行中的行程
      parameters:
        - $ref: '#/components/parameters/TripId'
      responses:
        '200':
          description: 行程已完成
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Trip' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    TripId: { name: tripId, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    Problem:
      description: RFC 9457问题详情
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    StartTripRequest:
      type: object
      additionalProperties: false
      required: [deviceId]
      properties:
        deviceId: { type: string, format: uuid }
      example: { deviceId: cb94d02f-3843-41a4-8312-b549690ba21d }
    Trip:
      type: object
      required: [id, deviceId, status, startedAt]
      properties:
        id: { type: string, format: uuid }
        deviceId: { type: string, format: uuid }
        status: { enum: [ACTIVE, COMPLETED, CANCELLED] }
        startedAt: { type: string, format: date-time }
        endedAt: { type: [string, 'null'], format: date-time }
    TrackPointBatchRequest:
      type: object
      additionalProperties: false
      required: [points]
      properties:
        points:
          type: array
          minItems: 1
          maxItems: 100
          items: { $ref: '#/components/schemas/TrackPointInput' }
      example:
        points:
          - { recordedAt: '2026-08-04T08:30:00Z', longitude: 116.397128, latitude: 39.916527, accuracyMeters: 8.5, speedMetersPerSecond: 1.1 }
    TrackPointInput:
      type: object
      additionalProperties: false
      required: [recordedAt, longitude, latitude, accuracyMeters]
      properties:
        recordedAt: { type: string, format: date-time }
        longitude: { type: number, minimum: -180, maximum: 180 }
        latitude: { type: number, minimum: -90, maximum: 90 }
        accuracyMeters: { type: number, minimum: 0, maximum: 1000 }
        speedMetersPerSecond: { type: [number, 'null'], minimum: 0 }
...
    Problem:
      type: object
      required: [status, code, title, detail, traceId, fieldErrors]
      properties:
        status: { type: integer }
        code: { type: string }
        title: { type: string }
        detail: { type: string }
        traceId: { type: string }
        fieldErrors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
      example: { status: 400, code: VALIDATION_FAILED, title: 请求参数校验失败, detail: 请求包含无效字段, traceId: 01J4K9Z8M8Q1, fieldErrors: [{ field: longitude, message: 最大值不能超过180 }] }
security:
  - bearerAuth: []